backend/requirements.txt
fastapi
uvicorn
openai-whisper
sentence-transformers
faiss-cpu
ffmpeg-python
pydantic
python-multipart
